xeno_V8.7 fixed publicExecute

+++ app/Http/Controllers/PublicExecuteCardController.php
@@ -40,15 +40,46 @@ class PublicExecuteCardController extends Controller
             $targetuser->card_2 = $drawCard->card_number;
             $targetuser->save();
             $drawCard->delete();
-            return redirect()->route('groups.show', [$group->id]);
+        } if ( !isset($targetuser->card_1) && isset($targetuser->card_2) ) {
+            $targetuser->card_1 = $drawCard->card_number;
+            $targetuser->save();
+            $drawCard->delete();
         }
-        $targetuser->card_1 = $drawCard->card_number;
-        $targetuser->save();
-        $drawCard->delete();
-
-        $group->publicexecutedcard_1 = $targetuser->card_1;
-        $group->publicexecutedcard_2 = $targetuser->card_2;
+        $group->publicexecutecard_1 = $targetuser->card_1;
+        $group->publicexecutecard_2 = $targetuser->card_2;
         $group->save();
+        return redirect()->route('groups.show', [$group->id])->with('message','処刑対象を選択してください。');;
+    }
+
+    public function publicExecutedLeftOrRightCard(Group $group,Request $request)// カード効果5疫病(リクエスト処理)
+    {
+        $group = Group::GroupWithUsersCardsDeadCardsKillCard($group)->first();
+        $targetuser = $group->users->where('publicexcutetarget', Auth::id())->first();
+
+        if( $request->publicexecute === 'left' ) {
+            $deadcard = new Deadcard;
+            $deadcard->group_id = $group->id;
+            $deadcard->card_number = $targetuser->card_1;
+            $deadcard->save();
+            $targetuser->card_1 = null;
+            $targetuser->publicexcutetarget = null;
+            $targetuser->save();
+            $group->publicexecutecard_1 = null;
+            $group->publicexecutecard_2 = null;
+            $group->save();
+        }
+        if ( $request->publicexecute === 'right' ) {
+            $deadcard = new Deadcard;
+            $deadcard->group_id = $group->id;
+            $deadcard->card_number = $targetuser->card_2;
+            $deadcard->save();
+            $targetuser->card_2 = null;
+            $targetuser->publicexcutetarget = null;
+            $targetuser->save();
+            $group->publicexecutecard_1 = null;
+            $group->publicexecutecard_2 = null;
+            $group->save();
+        }
         return redirect()->route('groups.show', [$group->id]);
     }
 }
